Models, Routes and Server Created
The first part is done, the crud application is working, I still need to fix the view all resources I own, but besides that, the crud part seems to be working fine. I am able to crud my resource. Next step is to fix the view all resources I own.